+++ packages/changed-elements-react/src/common/Widget/Widget.tsx
@@ -3,8 +3,8 @@
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
 import { forwardRef, type ReactElement } from "react";
-import { WidgetContent } from "./WidgetContent.js";
-import { WidgetTitleBar } from "./WidgetTitleBar.js";
+import { WidgetBody } from "./WidgetBody.js";
+import { WidgetHeader } from "./WidgetHeader.js";
 import { WidgetToolBar } from "./WidgetToolBar.js";
 
 import "./Widget.css";
@@ -18,39 +18,33 @@ export type WidgetProps = {
  * Widget component.
  * @example
  * <Widget>*
- *   <Widget.TitleBar>
+ *   <Widget.Header>
  *     My Widget title
- *   </Widget.TitleBar>
- *   <Widget.Content>
+ *   </Widget.Header>
+ *   <Widget.Body>
  *     Here is my Widget content
- *   </Widget.Content>
+ *   </Widget.Body>
  * </Widget>
  *
  * <Widget>
- *   <Widget.TitleBar>
- *     <Widget.TitleBar.Title>
+ *   <Widget.Header>
+ *     <Widget.Header.Label>
  *        My Widget title
- *     </Widget.TitleBar.Title>
- *     <Widget.TitleBar.Content>
- *       <IconButton
- *         size='small'
- *         styleType='borderless'
- *       >
+ *     </Widget.Header.Label>
+ *     <Widget.Header.Actions>
+ *       <IconButton size="small" styleType="borderless">
  *         <SvgClose />
  *       </IconButton>
- *     </Widget.TitleBar.Content>
- *   </Widget.TitleBar>
+ *     </Widget.Header.Actions>
+ *   </Widget.Header>
  *   <Widget.Toolbar>
- *     <IconButton
- *       size='small'
- *       styleType='borderless'
- *     >
+ *     <IconButton size="small" styleType="borderless">
  *      <SvgPlaceholder />
  *     </IconButton>
  *   </Widget.Toolbar>
- *   <Widget.Content>
+ *   <Widget.Body>
  *     Here is my Widget content
- *   </Widget.Content>
+ *   </Widget.Body>
  * </Widget>
  */
 
@@ -58,19 +52,14 @@ export const Widget = Object.assign(
   forwardRef<HTMLDivElement, WidgetProps>(function Widget(props, ref): ReactElement {
     const { children, className, style, ...rest } = props;
     return (
-      <div
-        ref={ref}
-        className={`itwin-common-widget ${className ?? ""}`}
-        style={{ ...style }}
-        {...rest}
-      >
+      <div ref={ref} className={`itwin-widget ${className ?? ""}`} style={{ ...style }} {...rest}>
         {children}
       </div>
     );
   }),
   {
-    TitleBar: WidgetTitleBar,
+    Header: WidgetHeader,
     ToolBar: WidgetToolBar,
-    Content: WidgetContent,
+    Body: WidgetBody,
   },
 );