+++ packages/changed-elements-react/CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased](https://github.com/iTwin/changed-elements-react/tree/HEAD/packages/changed-elements-react)
 
+### Minor changes
+
+* Update `ChangedElementsWidget` layout
+* Update search bar style in `ChangedElementsWidget`
+
 ### Fixes
 
 * Fix Elements Widget failing to restore its previous scroll offset when used as AppUI widget